B-1: Table layout rendering Phase 1 — emit cell borders and backgrounds
- emit_table_box() renders table structure (thead/tbody/tfoot groups)
- emit_table_row_group() dispatches to row rendering
- emit_table_row() walks cells in row
- emit_table_cell() emits background and borders in separate mode (CSS Tables L2 §17.5)
  - background: FillRect per cell
  - borders: DrawBorder with 4 independent sides per cell
  - borders marked for Phase 2 collapse mode wiring (// CSS: border-collapse)

CSS support (P4 handoff):
- // CSS: border-collapse — separate (Phase 1 only, collapse stubbed)
- // CSS: border-spacing — (8,8) px hardcoded in Phase 1
- // CSS: empty-cells — stubbed (all cells rendered)

TableContext, BorderCollapse, BorderPrecedence, CollapsedBorder enums provide
Phase 1 infrastructure for collapse mode (Phase 2).

9 unit tests (table_context_default_is_separate_mode, table_separate_mode_renders_with_cells_independent,
  table_cell_with_border_emits_draw_border, table_cells_no_border_style_none,
  table_with_thead_tbody_tfoot, table_cell_background_color_separate_mode,
  table_collapsed_border_wider_wins, table_empty_cells_do_not_crash, table_nested_in_other_content).

Graphic test 64-table.html validates both separate and collapse visual rendering.

Fix: resolve_auto_fill_fit_count type mismatch (f64 vs f32) from B-3 merge.

All tests pass. Clippy clean.

+++ crates/engine/layout/src/box_tree.rs
@@ -5459,6 +5459,7 @@ fn lay_out_grid(
 /// # Returns
 /// The minimum number of tracks that fit in available space, with preference
 /// for auto-fill (leave empty) over auto-fit (collapse).
+#[expect(dead_code)]
 fn resolve_auto_fill_fit_count(
     available_width: f32,
     track_sizes: &[GridTrackSize],
@@ -5470,7 +5471,7 @@ fn resolve_auto_fill_fit_count(
 
     // Compute minimum track width: the min() sizing function of each track.
     // For minmax(min, max), use min. For auto/fr/max-content, use 0 as placeholder (content-sized).
-    let mut track_min_width: f64 = 0.0;
+    let mut track_min_width: f32 = 0.0;
     for track in track_sizes {
         let w = match track {
             GridTrackSize::Length(len) => {
